requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.1
google-generativeai==0.4.1
gunicorn==21.2.0